Tentative RISC OS support.
There's no way to test this because as far as I know there's no version
of Python for RISC OS that's recent enough to have ctypes and actually
ships with a working one at this point.

Might be worth doing as a C extension, but the thought of trying to build
that on RISC OS is very upsetting to me.