Cleaner optional compilation.
The distutils documentation is an incomplete fucking mess. I went to the
source code to figure out how to subclass, and then looked at other
projects that would have had to deal with this crap as well; a lot of
them (like, for example, simplejson) seem to name their subclass
ve_build_ext (presumably after a common source), so I went with
convention.

Apparently Python 3 has an `optional' flag on Extension constructors. Fat
lot of good that does us now.